backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Azure SDKs
azure-identity>=1.17.0
azure-storage-blob>=12.22.0
azure-cosmos>=4.7.0
azure-ai-vision-imageanalysis>=1.0.0
openai>=1.40.0
azure-search-documents>=11.6.0

# Graph / pathfinding
gremlinpython>=3.7.0
networkx>=3.3

# Utilities
pydantic>=2.8.0
pydantic-settings>=2.4.0
httpx>=0.27.0
